pyproject.toml
[project]
name = "practice-fastapi"
version = "0.1.0"
description = "Practice FastAPI"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "psycopg>=3.2.9",
    "psycopg-binary>=3.2.9",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.5",
    "sqlalchemy>=2.0.41",
    "typer>=0.16.0",
    "uvicorn>=0.35.0",
]
